create : split upload video method process to sub method

+++ src/video/video.service.ts
@@ -15,55 +15,71 @@ export class VideoService {
         @Inject(STORAGE_TOKEN) private readonly storage: IStorage,
         private readonly configService: ConfigService,
         private readonly transcoderService: TranscoderService,
+        private readonly fileStorage: IFileStorage;
     ) { }
+
+    encodeVideo = async (file: Express.Multer.File,) => {
+        const storagePath = this.configService.get<string>('localStorage.localStoragePath')!;
+
+        await fs.ensureDir(storagePath);
+
+        const outputPath = join(storagePath, `encoded-${file.filename}.mp4`);
+
+        await this.transcoderService.transcode(file.path, outputPath);
+
+        return outputPath;
+    }
+
+    saveVideoMetadata = async (dto: CreateVideoDto, fileUrl: string) => {
+        const videoMetadata = {
+            ...dto,
+            fileUrl: fileUrl,
+        } as Partial<IVideo>;
+
+        await this.storage.insertItem<IVideo>('video', videoMetadata);
+        console.log('Video metadata saved successfully');
+    }
+
+    saveVideoFile = async (sourceFilePath: string) => {
+
+        await fs.pathExists(sourceFilePath);
+
+        const url = await this.fileStorage.saveFile(sourceFilePath);
+
+        return url;
+    }
+
+    uploadAndSaveVideo = async (file: Express.Multer.File, dto: CreateVideoDto) => {
+        const outputPath = await this.encodeVideo(file);
+
+        const url = await this.saveVideoFile(outputPath);
+
+        await this.saveVideoMetadata(dto, url);
+
+        await fs.remove(outputPath);
+    }
 
     getAllVideos = async () => {
         return this.storage.getItems<IVideo>('video');
     }
 
-    encodeAndSave = async (file: Express.Multer.File, dto: CreateVideoDto) => {
-        try {
-            const storagePath = this.configService.get<string>('localStorage.localStoragePath')!;
-
-            await fs.ensureDir(storagePath);
-            const outputPath = join(storagePath, `encoded-${file.filename}.mp4`);
-
-            await this.transcoderService.transcode(file.path, outputPath);
-            const videoMetadata = {
-                ...dto,
-                fileUrl: outputPath,
-            } as Partial<IVideo>;
-            await this.storage.insertItem<IVideo>('video', videoMetadata);
-            return videoMetadata;
-        } finally {
-            await fs.remove(file.path);
-        }
-    }
-
     updateVideo = async (dto: UpdateVideoDto, file?: Express.Multer.File) => {
         const existingVideo = await this.storage.getItem<IVideo>('video', dto.id);
+
         if (!existingVideo) throw new NotFoundException('Video not found');
 
         let updateData = { ...dto } as Partial<IVideo>;
 
         if (file) {
-            try {
-                const storagePath = this.configService.get<string>('localStorage.localStoragePath')!;
+            const outputPath = await this.encodeVideo(file);
 
-                await fs.ensureDir(storagePath);
-                const outputPath = join(storagePath, `encoded-${file.filename}.mp4`);
+            const url = await this.saveVideoFile(outputPath);
 
-                await this.transcoderService.transcode(file.path, outputPath);
+            if (existingVideo.fileUrl) { await fs.remove(existingVideo.fileUrl);}
+        
+            await fs.remove(outputPath);
 
-                if (existingVideo.fileUrl) {
-                    await fs.remove(existingVideo.fileUrl);
-                }
-                updateData['fileUrl'] = outputPath;
-            } catch (error) {
-                throw new InternalServerErrorException('Failed to update video');
-            } finally {
-                await fs.remove(file.path);
-            }
+            updateData['fileUrl'] = url;
         }
         return this.storage.updateItem<IVideo>('video', updateData);
     }
